Create crontrubutionCard component
